Rework modified classes for pipeline
Adjusted existing classes with changes from modified plugin version.
Classes which are too different from existing ones were placed into 'pipeline' package.
Existing classes might be adjusted to be compatible with pipeline, but not yet clear how to achieve that.
