+++ src/main/java/hudson/plugins/plot/PropertiesSeries.java
@@ -5,8 +5,10 @@
 
 package hudson.plugins.plot;
 
+import hudson.Extension;
 import hudson.FilePath;
 
+import hudson.model.Descriptor;
 import java.io.InputStream;
 import java.io.PrintStream;
 import java.util.ArrayList;
@@ -15,9 +17,11 @@ import java.util.Properties;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
+import net.sf.json.JSONObject;
 import org.apache.commons.io.IOUtils;
 import org.apache.commons.lang.ArrayUtils;
 import org.kohsuke.stapler.DataBoundConstructor;
+import org.kohsuke.stapler.StaplerRequest;
 
 /**
  * @author Allen Reese
@@ -75,6 +79,23 @@ public class PropertiesSeries extends Series {
             return null;
         } finally {
             IOUtils.closeQuietly(in);
+        }
+    }
+
+    @Override
+    public Descriptor<Series> getDescriptor() {
+        return new PropertiesSeries.DescriptorImpl();
+    }
+
+    @Extension
+    public static class DescriptorImpl extends Descriptor<Series> {
+        public String getDisplayName() {
+            return "";
+        }
+
+        @Override
+        public Series newInstance(StaplerRequest req, JSONObject formData) throws FormException {
+            return SeriesFactory.createSeries( formData, req );
         }
     }
 }
