+++ src/main/java/hudson/plugins/plot/Series.java
@@ -5,12 +5,17 @@
 
 package hudson.plugins.plot;
 
+import hudson.Extension;
 import hudson.FilePath;
 
+import hudson.model.AbstractDescribableImpl;
+import hudson.model.Descriptor;
 import java.io.PrintStream;
 import java.util.List;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
+import net.sf.json.JSONObject;
+import org.kohsuke.stapler.StaplerRequest;
 
 /**
  * Represents a plot data series configuration.
@@ -18,7 +23,7 @@ import java.util.regex.Pattern;
  * @author Nigel Daley
  * @author Allen Reese
  */
-public abstract class Series {
+public abstract class Series extends AbstractDescribableImpl<Series> {
     private static transient final Pattern PAT_NAME = Pattern.compile("%name%");
     private static transient final Pattern PAT_INDEX = Pattern
             .compile("%index%");
@@ -134,5 +139,22 @@ public abstract class Series {
         }
 
         return resultUrl;
+    }
+
+    @Override
+    public Descriptor<Series> getDescriptor() {
+        return new DescriptorImpl();
+    }
+
+    @Extension
+    public static class DescriptorImpl extends Descriptor<Series> {
+        public String getDisplayName() {
+            return "";
+        }
+
+        @Override
+        public Series newInstance(StaplerRequest req, JSONObject formData) throws FormException {
+            return SeriesFactory.createSeries(formData, req);
+        }
     }
 }
