+++ src/main/java/hudson/plugins/plot/pipeline/PlotPipelineDescriptor.java
@@ -4,6 +4,9 @@ import hudson.Extension;
 import hudson.FilePath;
 import hudson.model.AbstractProject;
 import hudson.model.Job;
+import hudson.plugins.plot.Messages;
+import hudson.plugins.plot.Plot;
+import hudson.plugins.plot.SeriesFactory;
 import hudson.tasks.BuildStepDescriptor;
 import hudson.tasks.Publisher;
 import hudson.util.FormValidation;
